タイマー作成時に、endTime指定をする機能の実装 #14 (#16)
* #14 utilsに一部移動

* #14 自動開始処理の実装
